Motor Backward Working

+++ Supervision/freeRTOS/Core/Inc/motors.h
@@ -2,7 +2,7 @@
 #include "stdint.h"
 
 void Motors_Init(TIM_HandleTypeDef *htim);
-void leftMotor(float value);
-void rightMotor(float value);
+void leftMotor(double value);
+void rightMotor(double value);
 double getLeftMotorValue();
 double getRightMotorValue();
